requirements.txt
altgraph==0.17.3
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
boto3==1.26.84
botocore==1.29.84
Brotli==1.0.9
bucketeer==0.1.8
captcha==0.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
colormap==1.0.4
contourpy==1.0.7
cryptography==39.0.2
cssselect2==0.7.0
cycler==0.11.0
distlib==0.3.6

django-crum==0.7.9
django-decouple==2.1
django-enumfield==3.1
django-fontawesome-5==1.0.18
django-phone-field==1.8.1
django-storages==1.13.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
easydev==0.12.1
ERP==0.27
filelock==3.12.0
fontawesome==5.10.1.post1
fontawesome-free==5.15.4
fontawesomefree==6.4.0
fonttools==4.39.4
gunicorn==20.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mysqlclient==2.1.1
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
pandas==2.0.1
pefile==2023.2.7
pexpect==4.8.0
Pillow==9.4.0
platformdirs==3.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pycryptodome==3.17
pydyf==0.6.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyMySQL==1.0.3
pyparsing==3.0.9
pypdf==3.5.2
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
reports==0.4.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.14
virtualenv==20.22.0
weasyprint==59.0
webencodings==0.5.1
whitenoise==6.4.0
win32-setctime==1.1.0
xhtml2pdf==0.2.9
zopfli==0.2.2
